% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsToUse.R
\name{colorsToUse}
\alias{colorsToUse}
\title{Color Palettes to Use}
\usage{
colorsToUse(palette, type = "Discreet")
}
\arguments{
\item{palette}{The color palette to use. Options are gigg.}

\item{type}{Discreet or Fade. Discreet is useful for categorical data. Fade is useful for cardinal data.}
}
\value{
Returns a dataframe with colorName, hex, and rgb.
}
\description{
Frequently used color palettes.
}
