% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDatasets.R
\name{compareDatasets}
\alias{compareDatasets}
\title{Compare columns of multiple datasets NOT WORKING YET}
\usage{
compareDatasets(targetDf, otherDfs)
}
\arguments{
\item{targetDf}{The name of the dataframe, in quotes, that you want to summarize.}

\item{otherDfs}{A vector of strings containing the names of other dataframes to compare.}
}
\value{
Dataframe.
}
\description{
A way to evaluate which columns from one dataset match up with columns from another dataset
}

