% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCallsToCharacter.R
\name{convColsToCharacter}
\alias{convColsToCharacter}
\title{Convert all columns to characters and convert NAs to blanks}
\usage{
convColsToCharacter(dataframe)
}
\arguments{
\item{dataframe}{The name of the dataframe, not in quotes, for which the columns need to be converted to character class.}
}
\value{
Dataframe
}
\description{
Converts all columns of a dataframe to character class, and also converts NAs to blanks.

I used to use this a lot, but not so much anymore.
}
