% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertClasses.R
\name{convertClasses}
\alias{convertClasses}
\title{Convert classes of columns from one dataset to classes of another}
\usage{
convertClasses(baseDf, dfToConvert)
}
\arguments{
\item{baseDf}{The dataframe with classes that will be used as the standard.}

\item{dfToConvert}{The dataframe with classes that will be converted to the same classes as the base.}
}
\value{
Dataframe.
}
\description{
Useful when trying to bind rows of a dataset that has the same column names of another dataset, but different classes.
}

