% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Creates a density plot of a quantitative vector}
\usage{
densityPlot(df, column, lowerx = -10, upperx = 10,
  title = "Density of Whatever You're Plotting")
}
\arguments{
\item{df}{The name of the dataframe, not in quotes.}

\item{column}{The name of the column, not in quotes, that will be graphed.}

\item{lowerx}{The lower bound of the x-axis to show on the chart. Default is -10.}

\item{upperx}{The upper bound of the x-axis to show on the chart. Default is 10.}

\item{title}{The title, in quotes, that you want to give to the plot. Default = "Density of Whatever You're Plotting".}
}
\value{
Densit plot.
}
\description{
This is useful for getting a quick idea of the distribution of a column of data. It also draws a line for the median.
}
