% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIt.R
\name{findIt}
\alias{findIt}
\title{Finds combinations of numbers that add up to a target value.}
\usage{
findIt(nums, target)
}
\arguments{
\item{nums}{A vector of numbers.}

\item{target}{The target number.}
}
\value{
A string with a list of numbers.
}
\description{
This was created to find a combination of check deposits that add up to a total deposit amount.
}
