% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeansCalcClusters.R
\name{kMeansCalcClusters}
\alias{kMeansCalcClusters}
\title{Function to mathematically calculate the optimal number of clusters}
\usage{
kMeansCalcClusters(data, nc = 40)
}
\arguments{
\item{data}{The name of the dataframe whose columns will be clustered.}

\item{nc}{The number of clusters to try. Defaults either to 40 or the number of rows in the dataframe.}
}
\value{
Number.
}
\description{
It should come close to calculating the elbow point mathematically for the optimal number of clusters using K-Means clustering.

It essentially draws a line from the least number of clusters to the max number of clusters and then calculates the point that's the farthest away from that line.

It assumes all columns are already numeric values.The columns used in the dataframe must already be numeric values, and they should already be standardized.

It's heavily dependent on the max number of clusters. A larger nc will have a higher accuracy.
}
