% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeansWssPlot.R
\name{kMeansWssPlot}
\alias{kMeansWssPlot}
\title{Returns an elbow plot of within group SS for a sequential number of K-means clusters}
\usage{
kMeansWssPlot(data, nc = 40)
}
\arguments{
\item{data}{The name of the dataframe whose columns will be clustered.}

\item{nc}{The number of clusters to try. Defaults either to 40 or the number of rows in the dataframe.}
}
\value{
Elbow plot.
}
\description{
Useful for eyeballing the elbow point for determining the optimal number of clusters.

The columns used in the dataframe must already be numeric values, and they should already be standardized.
}

