% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medReplace.R
\name{medReplace}
\alias{medReplace}
\title{Replace missing values of a column with the median}
\usage{
medReplace(x)
}
\arguments{
\item{x}{The name of the vector for which you want to replace missing values with the median.}
}
\value{
A number.
}
\description{
This is useful for data that has missing values and for which it makes sense to replace missing values with the median.

Especially useful when used with dplyr::group_by so that you can replace missing values with medians from the same strata.
}

