% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExcelAllSheets.R
\name{readExcelAllSheets}
\alias{readExcelAllSheets}
\title{Read in all Excel sheets from a single Excel file.}
\usage{
readExcelAllSheets(filename = "")
}
\arguments{
\item{filename}{The name of the Excel file, in quotes. The path should also be included if it's not in the working directory.}
}
\value{
Dataframes.
}
\description{
Each sheet will be saved as a separate dataframe using the name of the tab.
}
