% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Combine all files from a directory into one dataframe}
\usage{
readFiles(df)
}
\arguments{
\item{df}{The name of the dataframe that will be returned.}
}
\value{
Dataframe
}
\description{
Useful if you have a folder with many instances of essentially the same dataframe. This assumes that all of the files are tidy datasets.
}
