% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove0sd.R
\name{remove0sd}
\alias{remove0sd}
\title{Remove columns with standard deviation of zero}
\usage{
remove0sd(dataframe)
}
\arguments{
\item{dataframe}{The name of the dataframe}
}
\value{
Dataframe.
}
\description{
Useful for analyses that assume variation in every column of a dataframe.
}
