% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceColWithMode.R
\name{replaceColWithMode}
\alias{replaceColWithMode}
\title{Replace missing values of a column with the modal value}
\usage{
replaceColWithMode(df)
}
\arguments{
\item{df}{The name of the dataframe for which the columns will replace missing value with modal values}
}
\description{
This is useful for data that has missing values and for which it makes sense to replace missing values with the modal value.

Works with numeric or categorical columns

Especially useful when used with dplyr::group_by so that you can replace missing values with medians from the same strata.
}

