% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenStates.R
\name{shortenStates}
\alias{shortenStates}
\title{Converts full state name to two letter abbreviation}
\usage{
shortenStates(stateColumn)
}
\arguments{
\item{stateColumn}{The name of the column that has state names in it.}
}
\description{
Useful for messy data in which some rows contain full state names, and others contain the two letter abbreviation.
}

