% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsForecastDaily.R
\name{tsForecastDaily}
\alias{tsForecastDaily}
\title{Daily Time Series Forecast}
\usage{
tsForecastDaily(df, dateColumn, valueColumn, period = 28,
  seasonalPeriods = c(7, 364), K = 2, returnMePlot = F,
  returnYoyPlot = F)
}
\arguments{
\item{df}{The unquoted name of the dataframe that you want to summarize.}

\item{dateColumn}{The quoted name of the column that contains the daily dates.}

\item{valueColumn}{The quoted name of the column that has the daily values to be forecasted.}

\item{period}{The number of periods forecasted into the future.}

\item{seasonalPeriods}{The other periods, in addition to the period parameter, that may be influential}

\item{K}{The number of fourier terms. Must be one lesss than the number of periods}

\item{returnMePlot}{Return the model evaluation plot?}

\item{returnYoyPlot}{Return the year-over-year plot?}
}
\value{
List the contains a dataframe with the test, training, and forecasted data (dataFor), a dataframe with only the forecasted data (dataForOnly), mean absolute percentage error (mape), model evaluation plot (modelEvalPlot), and year over year dataframe including the forecast (yoySales).
}
\description{
Uses recursive partitioning and auto arima to make daily forecasts.

Requires the data to already be summarized into daily amounts
}
