% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsForecastHourly.R
\name{tsForecastHourly}
\alias{tsForecastHourly}
\title{Daily Time Series Forecast By Hour}
\usage{
tsForecastHourly(df, dateColumn, valueColumn, covs = NULL, algo = "rpart",
  lossFunction = "mape", period = 28, seasonalPeriods = c(7, 364),
  K = 2, returnMePlot = F, returnYoyPlot = F)
}
\arguments{
\item{df}{The unquoted name of the dataframe that you want to summarize.}

\item{dateColumn}{The quoted name of the column that contains the hourly dates. Should be in POSIXct format.}

\item{valueColumn}{The quoted name of the column that has the daily values to be forecasted.}

\item{covs}{Optional. A dataframe of covariates. This dataframe should include at least two columns: (1) A date column with the number of rows equal to the number of periods to forecast into the future, and (2) a column with values for each day.}

\item{algo}{Quoted name of algorithm to use. Defaults to rpart, which is fast. The other option is randomForest.}

\item{lossFunction}{Defaults to  "mape" (mean absolute percentage error). Other option is "mae" (mean absolute error).}

\item{period}{The number of periods forecasted into the future.}

\item{seasonalPeriods}{The other periods, in addition to the period parameter, that may be influential}

\item{K}{The number of fourier terms. Must be one lesss than the number of periods}

\item{returnMePlot}{Return the model evaluation plot? T = Yes.}

\item{returnYoyPlot}{Return the year-over-year plot? T = Yes.}
}
\value{
List the contains a dataframe with the test, training, and forecasted data (dataFor), a dataframe with only the forecasted data (dataForOnly), variable importance plot if randomForest is selected (viPlot), loss (either mape or mae), model evaluation plot (modelEvalPlot), and year over year dataframe including the forecast (yoySales).
}
\description{
Uses recursive partitioning or random forest and auto arima to make daily forecasts.

Requires the data to already be summarized into hourly amounts. Make sure that there are no missing periods of coviate data if covariates are being included in the model.
}
