% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterUpdate.R
\name{twitterUpdate}
\alias{twitterUpdate}
\title{Daily Twitter Data update}
\usage{
twitterUpdate(screenName, directory)
}
\arguments{
\item{screenName}{The screen name of the twitter account, without the hashtag. For example, RonGuymon or RickGuymon}

\item{directory}{Full directory path where the data will be stored. For example, /Volumes/GreatCustomer/Data/. If the directory doesn't already exist, a new one will be created.}
}
\value{
Returns a list with an update of what was written to file.
}
\description{
Uses the twitteR package to get witter data for: Timeline tweets, followers, following, hashtags, and mentions.

You need to authenticate before using this function:setup_twitter_oauth(twitter_api_key, twitter_api_secret, twitter_access_token, twitter_access_token_secret)

Run the function the same way regardless of whether you are updating your existing data, or if you are starting fresh.
}
