% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewMiss.R
\name{viewMiss}
\alias{viewMiss}
\title{View a simple histogram of missing data}
\usage{
viewMiss(df)
}
\arguments{
\item{df}{The name of the dataframe, not in quotes, for which the columns need to be converted to character class.}
}
\value{
Plot
}
\description{
Returns a histogram with a bar for each column in the dataframe. The height is the percent missing.
}

