% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopifyHourly.R
\name{shopifyHourly}
\alias{shopifyHourly}
\title{Get Shopify Data on an Hourly Basis}
\usage{
shopifyHourly(shopKey, shopPw, shopPath, hourSequence, consecutiveZeros)
}
\arguments{
\item{shopKey}{The quoted api key. Get this  and shopPw by creating a private app. Must use the customer's login credentials, to do so.}

\item{shopPw}{The quoted api password.}

\item{shopPath}{The quoted path. The path is something like "myStoreName.com"}

\item{hourSequence}{Vector of timestamps ordered from most recent to oldest. Doesn't have to be in hourly increments.}

\item{consecutiveZeros}{Integer representing the number of consecutive times in which no data is returned before stopping.}
}
\value{
A dataframe with a row for each item in the transaction. Some columns are summary for the whole order.
}
\description{
This function requires a Shopify Key, Password, and path.

The Shopify API key requires the store owner to go in and create it. I need to document how.



Depends on base64enc, httr, lubridate, tidyverse, magrittr, jsonlite.
}
